requirements.txt
altair==5.2.0
appdirs==1.4.4
argcomplete==3.2.3
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
b2==3.17.0
b2sdk==1.33.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.64
botocore==1.34.64
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
convertdate==2.4.0
cycler==0.12.1
Cython==3.0.9
debugpy==1.8.1
decorator==5.1.1
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.50.0
frozendict==2.4.0
gitdb==4.0.11
GitPython==3.1.42
html5lib==1.1
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
logfury==1.0.1
lxml==5.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multitasking==0.0.11
nest-asyncio==1.6.0
nselib==0.7
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
peewee==3.17.1
phx-class-registry==4.1.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.1
pydeck==0.8.1b0
Pygments==2.17.2
PyMeeus==0.5.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rst2ansi==0.1.5
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.32.2
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
upstox-python-sdk==2.4.1
urllib3==2.2.1
uuid==1.30
vaderSentiment==3.3.2
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
yfinance==0.2.37
